+++ package.json
@@ -35,7 +35,7 @@
     "@graasp/query-client": "0.1.5",
     "@graasp/sdk": "0.4.1",
     "@graasp/translations": "1.4.0",
-    "@graasp/ui": "https://github.com/graasp/graasp-ui.git#280-creative-commons",
+    "@graasp/ui": "0.11.1",
     "@mui/icons-material": "5.10.9",
     "@mui/lab": "5.0.0-alpha.104",
     "@mui/material": "5.10.12",
